refactor: organize tests by src structure
