+++ src/router.js
@@ -1,7 +1,7 @@
 import { createRouter, createWebHashHistory } from 'vue-router'
-import routes from '@/routes.js'
-import { useAuthentication } from '@/composables/useAuthentication.js'
-import { authEvents, AUTH_REQUIRED_EVENT } from '@/authEvents.js'
+import routes from './routes.js'
+import { useAuthentication } from './composables/useAuthentication.js'
+import { authEvents, AUTH_REQUIRED_EVENT } from './authEvents.js'
 
 /**
  * Create a Vue Router instance with a simple authentication guard.
